requirements.txt
asgiref==3.4.1
Django==3.0.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
Pillow==7.1.2
psycopg2==2.8.5
pytz==2021.3
sqlparse==0.4.2
